backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
langchain==0.1.16
langchain-openai==0.1.3
langchain-community==0.0.36
chromadb==0.4.24
openai==1.23.1
pymupdf==1.24.1
python-docx==1.1.0
python-pptx==0.6.23
sqlalchemy==2.0.29
pydantic==2.6.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
redis==5.0.3
